🔧 (gitignore): add _gen directory to .gitignore 📝 (capabilities): enhance capabilities page with detailed content and cards 📝 (book-online): update booking page with subtitle and button 💡 (render-link): improve readability of render-link.html 💡 (capabilities): refactor capabilities list and single templates for better structure 💡 (booking): improve readability of booking template ✨ (headline): add support for buttons in headline partial 💡 (header): improve readability of header partial 💡 (main-menu-section): improve readability and add overview link
Adding the _gen directory to .gitignore prevents generated files from being tracked. Enhancing the capabilities page provides more detailed information and a better user experience. Updating the booking page with a subtitle and button improves usability. Refactoring and improving readability in various templates ensures better maintainability and clarity in the codebase. Adding support for buttons in the headline partial allows for more interactive and informative headlines.

✨ (main-menu.html, our-customers.html): improve main menu formatting and enhance customer section layout

- Adjust indentation in `main-menu.html` for better readability.
- Change single quotes to double quotes in JavaScript for consistency.
- Add padding and text-center class to the container in `our-customers.html` for better styling.
- Increase the number of customer logos displayed from 12 to 18.
- Ensure customer logos are resized to 150x50 webp format for uniformity and faster loading.

These changes enhance the readability of the code, improve the visual layout of the customer section, and ensure consistent styling across the site.