📝 (content): update title in people section and add resourceDir in Hugo config
Change the title in the people section from "Who are we?" to "Our People" to provide a clearer and more direct description of the content. Add the `resourceDir` configuration in the Hugo config file to specify a directory for storing resource files, which helps in organizing and managing static resources more effectively.