✨ (content): add course learning experiences taxonomy and descriptions
Introduce a new taxonomy for course learning experiences to enhance
the categorization and discoverability of training courses. This
includes adding immersive, traditional, and guided learning
experiences to the Scrum Master course. New content files for
"Immersive" and "Traditional" learning experiences are created to
provide detailed descriptions. This change aims to improve the
user's ability to find courses that match their preferred learning
style, thereby enhancing the overall user experience on the site.