📝 (blog): update blog post with contributor and embed YouTube videos
Add Ralph Jocham as a contributor to acknowledge his input.
Replace raw YouTube links with Hugo's YouTube shortcode for better
embedding and maintainability. This enhances the readability and
presentation of the blog post.

🔧 (config): add edit URL for GitHub in hugo.yaml

Include an edit URL in the Hugo configuration to facilitate easy
content editing on GitHub. This encourages contributions and
improves the workflow for content updates.

✨ (layout): add GitHub contribution link to blog layout

Add a "Contribute on GitHub" link to the blog layout, allowing
readers to easily access and edit the content. This change
promotes community engagement and content accuracy.