🔧 (NKDAgility.com.sln): add Visual Studio solution file for project organization 📝 (baseof.html): integrate Trustpilot widget for enhanced customer feedback visibility ✨ (scrumorg-trustpilot.html): create partial for Trustpilot widget to modularize code 🔧 (staticwebapp.config.json): update configuration for networking and forwarding gateway
The addition of the Visual Studio solution file organizes the project
structure, facilitating easier management and development. The integration
of the Trustpilot widget in the base template and creation of a partial
for it enhances the website by displaying customer reviews, which can
improve trust and engagement. The updates to the static web app
configuration file ensure proper networking and security settings,
allowing only specific IP ranges and forwarded hosts, which enhances
security and performance.